requirements.txt
fastmcp>=0.1.0
pydantic>=2.0.0
newsapi-python>=0.2.7
